% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_anova_dose_time.R
\name{build_items_list}
\alias{build_items_list}
\title{This function computes the venn diagram of the genes associated to time, dose or their interaction}
\usage{
build_items_list(PvalMat, p.val.th = 0.01)
}
\arguments{
\item{PvalMat}{matrix with pvalue associated to the dose, timepoint and the dose*timepoint effect that is the output of the compute_anova_dose_time function}

\item{p.val.th}{is the threshold at which p.values are considered significant. Default = 0.01}
}
\value{
a list containing the genes in each position of the venn diagram
}
\description{
This function computes the venn diagram of the genes associated to time, dose or their interaction
}
\examples{
data("FC_WY14643")
exp_data = fc_data
pheno_data = pdata
PvalMat = compute_anova_dose_time(exp_data, pheno_data,dose_index = 2,time_point_index = 3)
ItemsList = build_items_list(PvalMat)

}
