% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hls_genes_clustering.R
\name{clustering_summary}
\alias{clustering_summary}
\title{This function evaluate the summary index for a clustering results by taking into account the within and between variability}
\usage{
clustering_summary(DB, cluster)
}
\arguments{
\item{DB}{your matrix dataset}

\item{cluster}{is a numeric vector of clustering results}
}
\value{
a vector of evaluation indexes
}
\description{
This function evaluate the summary index for a clustering results by taking into account the within and between variability
}
\keyword{clustering}
\keyword{evaluation}
