% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_anova_dose_time.R
\name{compute_anova_dose_time}
\alias{compute_anova_dose_time}
\title{This function computes a two way anova between dose and time for the expression value of every genes}
\usage{
compute_anova_dose_time(exp_data, pheno_data, dose_index, time_point_index)
}
\arguments{
\item{exp_data}{is the expression matrix with genes on the rows and samples on the columns}

\item{pheno_data}{is a dataframe with phenodata informations. Samples are on the rows. The columns should include the dose and time point information.}

\item{dose_index}{numeric value specifing the column of the phenodata table containing the doses}

\item{time_point_index}{numeric value specifing the column of the phenodata table containing the time points}
}
\value{
a matrix with pvalue associated to the dose, timepoint and the dose*timepoint effect
}
\description{
This function computes a two way anova between dose and time for the expression value of every genes
}
\examples{
data("FC_WY14643")
exp_data = fc_data
pheno_data = pdata
PvalMat = compute_anova_dose_time(exp_data, pheno_data,dose_index = 2,time_point_index = 3)

}
