% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd_TD.R
\name{compute_BMD_IC50}
\alias{compute_BMD_IC50}
\title{This function identify the BMD area and the IC50 value in the time and dose maps}
\usage{
compute_BMD_IC50(immy, coord, geneName, BMD_threshold = 0.58)
}
\arguments{
\item{immy}{z-maps of the fitted 3D model, with doses on the columns and time points on the rows}

\item{coord}{matrix with x and y coordinate. The first column contain the doses, while the second one the time points}

\item{geneName}{is a character string containing the gene name}

\item{BMD_threshold}{threshold defining the responsive gene value}
}
\value{
an object of class TinderMIX containing the fitted BMD object, the IC50 value. The function plot the map showing the responsive region.
}
\description{
This function identify the BMD area and the IC50 value in the time and dose maps
}
