% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{compute_enrichment}
\alias{compute_enrichment}
\title{This function perform enchment of the genes in each cluster}
\usage{
compute_enrichment(optimal_clustering, corrType = "fdr",
  type_enrich = "KEGG", org_enrich = "rnorvegicus", pth = 0.05,
  sig = FALSE, mis = 0, only_annotated = FALSE)
}
\arguments{
\item{optimal_clustering}{vector of final clustering}

\item{corrType}{string specifing the algorithm used for determining the significance threshold, one of gSCS, fdr, bonferroni. Default: fdr}

\item{type_enrich}{string specifying the enrichment type. Default = KEGG}

\item{org_enrich}{string specifying the organism. Default = rnorvegicus}

\item{pth}{numeric value specifyint the pvalue threshold. Default = 0.05}

\item{sig}{whether all or only statistically significant results should be returned}

\item{mis}{minimum size of functional category, smaller categories are excluded}

\item{only_annotated}{statistical domain size, one of "annotated", "known"}
}
\value{
a list with the enriched pathways for each cluster of genes
}
\description{
This function perform enchment of the genes in each cluster
}
\examples{
  data("WY14643")
  exp_data = WY14643$exp_data
  pheno_data = WY14643$pheno_data
  PvalMat = compute_anova_dose_time(exp_data, pheno_data,dose_index = 2,time_point_index = 3)
  ItemsList = build_items_list(PvalMat)
  responsive_genes = unique(c(unlist(ItemsList$Dose),unlist(ItemsList$Time),unlist(ItemsList$`Dose:Time:DoseTime`),unlist(ItemsList$`Dose:Time`)))
  contour_res = create_contour(exp_data, pheno_data, responsive_genes,dose_index = 2,time_point_index =3 ,gridSize = 50)
  hls_res = hls_genes_clustering(contour_res$GenesMap,  nClust = c(5,10,15,20,25), method="pearson", hls.method = "ward")
  clpr = create_prototypes(clust_res = hls_res,summaryMat = hls_res$summaryMat,contour_res )
  enrRes = compute_enrichment(clpr$optcl,corrType = "fdr",type_enrich="KEGG", org_enrich = "rnorvegicus",pth = 0.05,sig = FALSE,mis = 0,only_annotated=FALSE)

}
