% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pairwise_fold_changes.R
\name{compute_fc}
\alias{compute_fc}
\title{This function starts from a phenodata and gene expression data matrix and compute all the possible pairwise foldchange values}
\usage{
compute_fc(exp_data, pheno_data, dose_index, time_index)
}
\arguments{
\item{exp_data}{is the expression matrix with genes on the rows and samples on the columns}

\item{pheno_data}{is a dataframe with phenodata informations. Samples are on the rows. The columns should include the dose and time point information. Doses of controls need to be indicated as 0}

\item{dose_index}{index of the column containiing the dose}

\item{time_index}{index of the column containing the time}
}
\value{
a list containing two new matrices
\item{fc_data}{a matrix with all the possible fold_changes}
\item{pdata}{the new phenodata table}
}
\description{
This function starts from a phenodata and gene expression data matrix and compute all the possible pairwise foldchange values
}
