% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{compute_pathways}
\alias{compute_pathways}
\title{This function perform enchment of a set of genes}
\usage{
compute_pathways(geneList = rownames(res$Mat), corrType = "fdr",
  type_enrich = "KEGG", annType = "SYMBOL", org_enrich = "hsapiens",
  pth = 0.05, sig = FALSE, mis = 0, only_annotated = FALSE)
}
\arguments{
\item{geneList}{vector of gene identifiers}

\item{corrType}{string specifing the algorithm used for determining the significance threshold, one of gSCS, fdr, bonferroni. Default: fdr}

\item{type_enrich}{string specifying the enrichment type. Default = KEGG}

\item{annType}{gene annotation type. default = SYMBOL}

\item{org_enrich}{string specifying the organism. Default = rnorvegicus}

\item{pth}{numeric value specifyint the pvalue threshold. Default = 0.05}

\item{sig}{whether all or only statistically significant results should be returned}

\item{mis}{minimum size of functional category, smaller categories are excluded}

\item{only_annotated}{statistical domain size, one of "annotated", "known"}
}
\value{
a list with the enriched pathways for each cluster of genes
}
\description{
This function perform enchment of a set of genes
}
