% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{convert_genes}
\alias{convert_genes}
\title{This function convert genes identifiers}
\usage{
convert_genes(organism = "hsapiens", GList, annType = "SYMBOL")
}
\arguments{
\item{organism}{a string specifying the organism under analysis}

\item{GList}{a list of genes identifier}

\item{annType}{string specifying the wanted gene identifier}
}
\value{
a list with the converted genes identifiers
}
\description{
This function convert genes identifiers
}
