% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{create_pathway_prototypes}
\alias{create_pathway_prototypes}
\title{This function create prototypes for a list of pathways}
\usage{
create_pathway_prototypes(enrichedPath = enrichedPath, annIDs,
  contour_res = contour_res, nPerm = 100)
}
\arguments{
\item{enrichedPath}{dataframe of enriched pathways coming from the compute_pathways function}

\item{annIDs}{vector with the pathways IDs for which the prototype will be computed}

\item{contour_res}{a list with the contours object returned in output by the create_contour function}

\item{nPerm}{number of permutation to run to compute the pvalue associated to pathway correaltion}
}
\value{
a list with the prototypes of the pathways, their genes correlation and a pvalue
}
\description{
This function create prototypes for a list of pathways
}
