% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hls_genes_clustering.R
\name{create_prototypes}
\alias{create_prototypes}
\title{this function create the cluster prototypes as the mean values of the z maps of all the genes in the cluster}
\usage{
create_prototypes(clust_res, summaryMat, contour_res)
}
\arguments{
\item{clust_res}{the clustering results object given in input by the function hls_genes_clustering}

\item{summaryMat}{the matrix with the summary statistics computed for the different k values}

\item{contour_res}{a list with the contours object returned in output by the create_contour function}
}
\value{
a list with the contour objects for each clustering prototype and the vector of the optimal clustering
\item{meanXYZ}{a list the contour object for each prototype}
\item{optcl}{a vector with the optimal clustering}
}
\description{
this function create the cluster prototypes as the mean values of the z maps of all the genes in the cluster
}
\examples{
  data("WY14643")
  exp_data = WY14643$exp_data
  pheno_data = WY14643$pheno_data
  PvalMat = compute_anova_dose_time(exp_data, pheno_data,dose_index = 2,time_point_index = 3)
  ItemsList = build_items_list(PvalMat)
  responsive_genes = unique(c(unlist(ItemsList$Dose),unlist(ItemsList$Time),unlist(ItemsList$`Dose:Time:DoseTime`),unlist(ItemsList$`Dose:Time`)))
  contour_res = create_contour(exp_data, pheno_data, responsive_genes,dose_index = 2,time_point_index =3 ,gridSize = 50)
  hls_res = hls_genes_clustering(contour_res$GenesMap,  nClust = c(5,10,15,20,25), method="pearson", hls.method = "ward")
  clpr = create_prototypes(clust_res = hls_res,summaryMat = hls_res$summaryMat,contour_res )

}
