% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_label.R
\name{create_tic_tac_toe}
\alias{create_tic_tac_toe}
\title{This function assigns a label to contour map}
\usage{
create_tic_tac_toe(map = t(clpr$meanXYZ[[1]][[3]]))
}
\arguments{
\item{map}{matrix containing the z-map for a specific gene or cluster prototype}
}
\value{
a list with 9x9 matrices specifying if the gene is active at low, mid or high time points and dose levels
}
\description{
This function assigns a label to contour map
}
