% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{create_tic_tac_toe_wordcloud}
\alias{create_tic_tac_toe_wordcloud}
\title{This function perform enrichment for each gene label and compute a pathway wordcloud for every label}
\usage{
create_tic_tac_toe_wordcloud(Enriched_list = Enriched_list,
  max.words = 200, scale = c(0.8, 2.5), random.order = FALSE,
  min.freq = 0)
}
\arguments{
\item{max.words}{max number of words in wordcluds}

\item{scale}{a vector of length 2 indicating the range of the size of the words.}

\item{random.order}{plot words in random order. If false, they will be plotted in decreasing frequency}

\item{min.freq}{words with frequency below min.freq will not be plotted}

\item{Mat}{matrix of gene labels}
}
\value{
a list with the enriched pathways for each cluster of genes
}
\description{
This function perform enrichment for each gene label and compute a pathway wordcloud for every label
}
