% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrich}
\alias{enrich}
\title{This function perform enchment of the genes in each cluster}
\usage{
enrich(x, type, org, pval, adjust_method, sig = FALSE, mis = 0,
  only_annotated = TRUE)
}
\arguments{
\item{x}{a dataframe with the gene names on the first column}

\item{type}{string specifying the enrichment type. Default = KEGG}

\item{org}{string specifying the organism. Default = rnorvegicus}

\item{pval}{numeric value specifyint the pvalue threshold. Default = 0.05}

\item{adjust_method}{string specifing the algorithm used for determining the significance threshold, one of gSCS, fdr, bonferroni. Default: fdr}

\item{sig}{whether all or only statistically significant results should be returned}

\item{mis}{minimum size of functional category, smaller categories are excluded}

\item{only_annotated}{statistical domain size, one of "annotated", "known"}
}
\value{
a list with the enriched pathways for each cluster of genes
}
\description{
This function perform enchment of the genes in each cluster
}
