% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hls_genes_clustering.R
\name{findCenter}
\alias{findCenter}
\title{This function select cluster prototypes. It select the prototype as the more correlated with the other elements in the cluster}
\usage{
findCenter(DB, clust_vector)
}
\arguments{
\item{DB}{your matrix dataset}

\item{clust_vector}{a numeric vector of clustering results}
}
\value{
a matrix of prototypes
}
\description{
This function select cluster prototypes. It select the prototype as the more correlated with the other elements in the cluster
}
\keyword{clustering;}
\keyword{correlation}
\keyword{multi-view}
\keyword{prototype;}
