% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_test.R
\name{fisher_test}
\alias{fisher_test}
\title{This function construct confusion matrix between patient classes and the obtained clustering}
\usage{
fisher_test(classes, clustering, matrixRownames, nCluster)
}
\arguments{
\item{classes}{is a vector of patient labels}

\item{clustering}{is a vector of clustering results}

\item{matrixRownames}{is a vector of names to assign as rownames of the confusion matrix}

\item{nCluster}{is the number of obtained clusters}
}
\value{
the confusion matrix
}
\description{
This function construct confusion matrix between patient classes and the obtained clustering
}
\keyword{clustering;}
\keyword{confusion}
\keyword{matrix}
\keyword{multi-view}
