% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd_TD.R
\name{label2DMap}
\alias{label2DMap}
\title{This function assigns a label to contour map}
\usage{
label2DMap(map, th = 0.95, nDoseInt = 3, nTimeInt = 3,
  doseLabels = c("Low", "Mid", "High"), timeLabels = c("High", "Mid",
  "Low"))
}
\arguments{
\item{map}{matrix containing the z-map for a specific gene or cluster prototype}

\item{th}{a threshold to define the portion of dose-response area to be identified as labels for the gene.}

\item{nDoseInt}{number of dose related breaks in the gene label's table. default is 3}

\item{nTimeInt}{number of time related breaks in the gene label's table. default is 3}

\item{doseLabels}{vector of colnames (doses) for the gene label's table. default is  c("Low","Mid","High")}

\item{timeLabels}{vector of rownames (time points) for the gene label's table. default c("High","Mid","Low")}
}
\value{
a list with 9x9 matrices specifying if the gene is active at low, mid or high time points and dose levels
}
\description{
This function assigns a label to contour map
}
\examples{
data("FC_WY14643")
exp_data = fc_data
pheno_data = pdata
  PvalMat = compute_anova_dose_time(exp_data, pheno_data,dose_index = 2,time_point_index = 3)
  ItemsList = build_items_list(PvalMat)
  responsive_genes = unique(c(unlist(ItemsList$Dose),unlist(ItemsList$Time),unlist(ItemsList$`Dose:Time:DoseTime`),unlist(ItemsList$`Dose:Time`)))
  contour_res = create_contour(exp_data, pheno_data, responsive_genes,dose_index = 2,time_point_index =3 ,gridSize = 50)
  hls_res = hls_genes_clustering(contour_res$GenesMap,  nClust = c(5,10,15,20,25), method="pearson", hls.method = "ward")
  clpr = create_prototypes(clust_res = hls_res,summaryMat = hls_res$summaryMat,contour_res )
  create_tic_tac_toe(map = t(clpr$meanXYZ[[1]][[3]]))
}
