% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_contour.R
\name{plot3d}
\alias{plot3d}
\title{This function plots the fitted 3d surface for the expression value of a gene}
\usage{
plot3d(toPlot = list(x, y, z), DF)
}
\arguments{
\item{toPlot}{is a list containing the predicted value for the x, y and z axis}

\item{DF}{is the data frame containing the information for the samples used in the fitting process}
}
\value{
a plotly object
}
\description{
This function plots the fitted 3d surface for the expression value of a gene
}
\examples{
  data("WY14643")
  exp_data = WY14643$exp_data
  pheno_data = WY14643$pheno_data
  PvalMat = compute_anova_dose_time(exp_data, pheno_data,dose_index = 2,time_point_index = 3)
  ItemsList = build_items_list(PvalMat)
  responsive_genes = unique(c(unlist(ItemsList$Dose),unlist(ItemsList$Time),unlist(ItemsList$`Dose:Time:DoseTime`),unlist(ItemsList$`Dose:Time`)))
  contour_res = create_contour(exp_data, pheno_data, responsive_genes,dose_index = 2,time_point_index =3 ,gridSize = 50)
  plot3d(toPlot = contour_res$RPGenes[["Pdk4"]],DF = contour_res$DFList[["Pdk4"]])

}
