% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_cake_diagrams_time_dose_effect}
\alias{plot_cake_diagrams_time_dose_effect}
\title{this function takes in input the result of the function run_all_BMD_IC50 and plot a 3x3 multiplot with the number of dose responsive genes fir tge 12 segment of time and dose interaction
the letters d and t (independently if they are capital or small ) stand for dose and time
+/- indicate if the gene fc is increasing or decreasing with respect of dose and time
capital letters are used to indicate which between dose and time has a stronger effect}
\usage{
plot_cake_diagrams_time_dose_effect(res)
}
\arguments{
\item{res}{is the result object from the run_all_BMD_IC50 function}
}
\value{
a ggplot object
}
\description{
this function takes in input the result of the function run_all_BMD_IC50 and plot a 3x3 multiplot with the number of dose responsive genes fir tge 12 segment of time and dose interaction
the letters d and t (independently if they are capital or small ) stand for dose and time
+/- indicate if the gene fc is increasing or decreasing with respect of dose and time
capital letters are used to indicate which between dose and time has a stronger effect
}
