% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hls_genes_clustering.R
\name{plot_clusters_prototypes}
\alias{plot_clusters_prototypes}
\title{this function plots the clusters prototype}
\usage{
plot_clusters_prototypes(meanXYZ, nR = 2, nC = 5,
  activity_threshold = activity_threshold,
  BMD_resonse_threhold = BMD_resonse_threhold, nDoseInt = nDoseInt,
  nTimeInt = nTimeInt, doseLabels = doseLabels,
  timeLabels = timeLabels)
}
\arguments{
\item{meanXYZ}{a list the contour object for each prototype computed with the function create_prototypes}

\item{nR}{the number of rows to use in the plot}

\item{contour_size}{number specifying the distance between each contour line}
}
\description{
this function plots the clusters prototype
}
