% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dynamic_dose_responsive_map.R
\name{plot_dynamic_dose_responsive_map}
\alias{plot_dynamic_dose_responsive_map}
\title{This function takes in input the result of the function create_contour and plot the dynamic dose responsive activation map of a specific gene}
\usage{
plot_dynamic_dose_responsive_map(contour_res, geneName, activity_threshold,
  BMD_response_threshold, mode, nTimeInt, nDoseInt, timeLabels, doseLabels)
}
\arguments{
\item{contour_res}{is the result object from the create_contour function}

\item{geneName}{is the name of the gene}
}
\value{
a ggplot object
}
\description{
This function takes in input the result of the function create_contour and plot the dynamic dose responsive activation map of a specific gene
}
