% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_kegg_radar_chart}
\alias{plot_kegg_radar_chart}
\title{this function takes in input the pathways enriched and retur a radar chart for each one of the gene label category}
\usage{
plot_kegg_radar_chart(Enriched_list, n = 5, vlcex = 1.5,
  kegg_level = 1, mar = c(2, 1, 1, 1))
}
\arguments{
\item{Enriched_list}{is the list of dataframe resulting from enrichment for each gene category}

\item{n}{is the max number of pathways to plot in each radar plot}

\item{vlcex}{is the size of the labels}

\item{kegg_level}{is the level of the kegg hierarchy to be considered in the plotting}

\item{mar}{are the margin settings for the plot}

\item{caxislabels}{is a character vector for center axis labels, overwriting values specified in axistype option. If NULL, the values specified by axistype option are used. Default is NULL.}
}
\value{
a ggplot object
}
\description{
this function takes in input the pathways enriched and retur a radar chart for each one of the gene label category
}
