% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_number_genes_labels}
\alias{plot_number_genes_labels}
\title{This function takes in input the result of the function run_all_BMD_IC50 and plot a 3x3 
heatmap with the number of dose responsive genes for each label}
\usage{
plot_number_genes_labels(res, drugName)
}
\arguments{
\item{res}{is the result object from the run_all_BMD_IC50 function}

\item{drugName}{is the name of the drug that will be used in the title}
}
\value{
a ggplot object
}
\description{
This function takes in input the result of the function run_all_BMD_IC50 and plot a 3x3 
heatmap with the number of dose responsive genes for each label
}
