% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_allsheets.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{read excel file as a list of dataframe}
\usage{
read_excel_allsheets(filename, tibble = FALSE)
}
\arguments{
\item{filename}{is the path to the file}

\item{tibble}{boolean specifying if the content of each sheet should be read as tibble or dataframe}
}
\value{
a data frame contained in the excel file
}
\description{
read excel file as a list of dataframe
}
