% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{write_xlsx_for_funmappone}
\alias{write_xlsx_for_funmappone}
\title{This function create an excel file with the same format of the input need by the FunMappOne tool}
\usage{
write_xlsx_for_funmappone(optimal_clustering,
  filePath = "../contour_clustering/gene_clustering.xlsx")
}
\arguments{
\item{optimal_clustering}{is a numeric vector with the clustering result for every gene}

\item{filePath}{is a string specifying the path of the xlsx file}
}
\description{
This function create an excel file with the same format of the input need by the FunMappOne tool
}
\examples{
  data("WY14643")
  exp_data = WY14643$exp_data
  pheno_data = WY14643$pheno_data
  PvalMat = compute_anova_dose_time(exp_data, pheno_data,dose_index = 2,time_point_index = 3)
  ItemsList = build_items_list(PvalMat)
  responsive_genes = unique(c(unlist(ItemsList$Dose),unlist(ItemsList$Time),unlist(ItemsList$`Dose:Time:DoseTime`),unlist(ItemsList$`Dose:Time`)))
  contour_res = create_contour(exp_data, pheno_data, responsive_genes,dose_index = 2,time_point_index =3 ,gridSize = 50)
  hls_res = hls_genes_clustering(contour_res$GenesMap,  nClust = c(5,10,15,20,25), method="pearson", hls.method = "ward")
  enrRes = compute_enrichment(clpr$optcl,corrType = "fdr",type_enrich="KEGG", org_enrich = "rnorvegicus",pth = 0.05,sig = FALSE,mis = 0,only_annotated=FALSE)
  write_xlsx_for_funmappone(clpr$optcl,filePath = "../contour_clustering/gene_clustering.xlsx")

}
