% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd_TD.R
\name{compute_BMD_IC50}
\alias{compute_BMD_IC50}
\title{This function identify the BMD area and the IC50 value in the time and dose maps}
\usage{
compute_BMD_IC50(immy, coord, geneName, activity_threshold = 0.1,
  BMD_response_threshold = 0.95, nDoseInt = 3, nTimeInt = 3,
  doseLabels = c("Late", "Middle", "Early"),
  timeLabels = c("Sensitive", "Intermediate", "Resilient"),
  toPlot = TRUE, addLegend = TRUE, tosave = FALSE, path = ".",
  mode = "cumulative")
}
\arguments{
\item{immy}{z-maps of the fitted 3D model, with doses on the columns and time points on the rows}

\item{coord}{matrix with x and y coordinate. The first column contain the doses, while the second one the time points}

\item{geneName}{is a character string containing the gene name}

\item{activity_threshold}{threshold defining the responsive gene area. Eg. if the immy maps contains genes logFC, then an activity_threhdold = 0.58 means that the active area will be the one with an effect of 1.5 bigger or smaller than the controls}

\item{BMD_response_threshold}{a threshold to define the portion of dose-response area to be identified as labels for the gene.}

\item{nDoseInt}{number of dose related breaks in the gene label's table. default is 3}

\item{nTimeInt}{number of time related breaks in the gene label's table. default is 3}

\item{doseLabels}{vector of colnames (doses) for the gene label's table. default is  c("Sensitive","Intermediate","Resilient")}

\item{timeLabels}{vector of rownames (time points) for the gene label's table. default c("Late","Middle","Early")}

\item{toPlot}{it true the gene map is displayed}

\item{addLegend}{if true the legend will be added to the plot}

\item{tosave}{if true a png of the gene map is saved in path}

\item{path}{path of the folder where to save the gene map}

\item{mode}{is a character specifying when an area is called active. values can be "cumulative" or "presence". If presence, an area is called active if at least one of its pixel is on the BMD curve. If cumulative, the number of region needed to reach the th% of the cumulative of the number of pixel on the BMD curve is identified.}
}
\value{
an object of class TinderMIX containing the fitted BMD object, the IC50 value. The function plot the map showing the responsive region.
}
\description{
This function identify the BMD area and the IC50 value in the time and dose maps
}
