% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{compute_enrichment}
\alias{compute_enrichment}
\title{This function perform enchment of the genes in each cluster}
\usage{
compute_enrichment(optimal_clustering, corrType = "fdr",
  type_enrich = "KEGG", org_enrich = "rnorvegicus", pth = 0.05,
  sig = FALSE, mis = 0, only_annotated = FALSE)
}
\arguments{
\item{optimal_clustering}{vector of final clustering}

\item{corrType}{string specifing the algorithm used for determining the significance threshold, one of gSCS, fdr, bonferroni. Default: fdr}

\item{type_enrich}{string specifying the enrichment type. Default = KEGG}

\item{org_enrich}{string specifying the organism. Default = rnorvegicus}

\item{pth}{numeric value specifyint the pvalue threshold. Default = 0.05}

\item{sig}{whether all or only statistically significant results should be returned}

\item{mis}{minimum size of functional category, smaller categories are excluded}

\item{only_annotated}{statistical domain size, one of "annotated", "known"}
}
\value{
a list with the enriched pathways for each cluster of genes
}
\description{
This function perform enchment of the genes in each cluster
}
