% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hls_genes_clustering.R
\name{create_prototypes}
\alias{create_prototypes}
\title{this function create the cluster prototypes as the mean values of the z maps of all the genes in the cluster}
\usage{
create_prototypes(clust_res, contour_res, optcl, mode = "mean")
}
\arguments{
\item{clust_res}{the clustering results object given in input by the function hls_genes_clustering}

\item{contour_res}{a list with the contours object returned in output by the create_contour function}

\item{optcl}{a vector with the optimal clustering}

\item{mode}{string specifying the kind of prototype to be estimated. Possible values are mean, median or prot. mean and median compute the prototype as the mean of median values of the gene maps falling in the same cluster. Prot compute the prototype as the map of the most central gene in the cluster. default is mean}

\item{summaryMat}{the matrix with the summary statistics computed for the different k values}
}
\value{
a list, named meanXYZ,  with the contour objects for each clustering prototype
}
\description{
this function create the cluster prototypes as the mean values of the z maps of all the genes in the cluster
}
