% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_label.R
\name{create_tic_tac_toe}
\alias{create_tic_tac_toe}
\title{This function assigns a label to contour map}
\usage{
create_tic_tac_toe(map = t(clpr$meanXYZ[[1]][[3]]))
}
\arguments{
\item{map}{matrix containing the z-map for a specific gene or cluster prototype}
}
\value{
a list with 9x9 matrices specifying if the gene is active at low, mid or high time points and dose levels
}
\description{
This function assigns a label to contour map
}
\examples{
data("FC_WY14643")
exp_data = fc_data
pheno_data = pdata
  PvalMat = compute_anova_dose_time(exp_data, pheno_data,dose_index = 2,time_point_index = 3)
  ItemsList = build_items_list(PvalMat)
  responsive_genes = unique(c(unlist(ItemsList$Dose),unlist(ItemsList$Time),unlist(ItemsList$`Dose:Time:DoseTime`),unlist(ItemsList$`Dose:Time`)))
  contour_res = create_contour(exp_data, pheno_data, responsive_genes,dose_index = 2,time_point_index =3 ,gridSize = 50)
  hls_res = hls_genes_clustering(contour_res$GenesMap,  nClust = c(5,10,15,20,25), method="pearson", hls.method = "ward")
  clpr = create_prototypes(clust_res = hls_res,summaryMat = hls_res$summaryMat,contour_res )
  create_tic_tac_toe(map = t(clpr$meanXYZ[[1]][[3]]))
}
