% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hls_genes_clustering.R
\name{hls_genes_clustering}
\alias{hls_genes_clustering}
\title{It clusters the contour plots by also estimating the ideal number of clusters}
\usage{
hls_genes_clustering(GenesMap, nClust = c(5, 10, 25, 50, 75, 100, 125,
  150, 175, 200, 250, 300), method = "pearson", hls.method = "ward")
}
\arguments{
\item{GenesMap}{a matrix with the z-maps computed with the function create_contour}

\item{nClust}{vector of putative numbers used to determine number of clusters. Default = c(5,10,25,50,75,100,125,150,175,200,250,300)}

\item{method}{string specifying the correlation method. Default = "pearson"}

\item{hls.method}{string specifying method for the hierarchical clustering. Default = "ward"}
}
\value{
a list with the clustering results and statistics for the optimal number of clusters
\item{hls_res}{a list containing the clustering results, the clustering vector and the centers for each k value given in input}
\item{summaryMat}{a matrix with summary statistic for each k value given in input}
\item{clusterList}{a list with the clustering vectors}
}
\description{
It clusters the contour plots by also estimating the ideal number of clusters
}
\examples{
data("FC_WY14643")
exp_data = fc_data
pheno_data = pdata
  PvalMat = compute_anova_dose_time(exp_data, pheno_data,dose_index = 2,time_point_index = 3)
  ItemsList = build_items_list(PvalMat)
  responsive_genes = unique(c(unlist(ItemsList$Dose),unlist(ItemsList$Time),unlist(ItemsList$`Dose:Time:DoseTime`),unlist(ItemsList$`Dose:Time`)))
  contour_res = create_contour(exp_data, pheno_data, responsive_genes,dose_index = 2,time_point_index =3 ,gridSize = 50)
  hls_res = hls_genes_clustering(contour_res$GenesMap,  nClust = c(5,10,15,20,25), method="pearson", hls.method = "ward")

}
