% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd_TD.R
\name{label2DMap}
\alias{label2DMap}
\title{This function assigns a label to contour map}
\usage{
label2DMap(map, coord, myContour, th = 0.95, mode = "cumulative",
  nDoseInt = 3, nTimeInt = 3, doseLabels = c("Late", "Middle",
  "Early"), timeLabels = c("Sensitive", "Intermediate", "Resilient"),
  toplot = FALSE)
}
\arguments{
\item{map}{matrix containing the z-map for a specific gene or cluster prototype}

\item{coord}{matrix with x and y coordinate. The first column contain the doses, while the second one the time points}

\item{myContour}{matrix with coordinate of bmd area border}

\item{th}{a threshold to define the portion of dose-response area to be identified as labels for the gene.}

\item{mode}{is a character specifying when an area is called active. values can be "cumulative" or "presence". If presence, an area is called active if at least one of its pixel is on the BMD curve. If cumulative, the number of region needed to reach the th% of the cumulative of the number of pixel on the BMD curve is identified.}

\item{nDoseInt}{number of dose related breaks in the gene label's table. default is 3}

\item{nTimeInt}{number of time related breaks in the gene label's table. default is 3}

\item{doseLabels}{vector of colnames (doses) for the gene label's table. default is  c("Sensitive","Intermediate","Resilient")}

\item{timeLabels}{vector of rownames (time points) for the gene label's table. default c("Late","Middle","Early")}

\item{nDoseInt}{number of dose related breaks in the gene label's table. default is 3}

\item{nTimeInt}{number of time related breaks in the gene label's table. default is 3}

\item{doseLabels}{vector of colnames (doses) for the gene label's table. default is  c("Sensitive","Intermediate","Resilient")}

\item{timeLabels}{vector of rownames (time points) for the gene label's table. default c("Late","Middle","Early")}
}
\value{
a list with 9x9 matrices specifying if the gene is active at low, mid or high time points and dose levels
}
\description{
This function assigns a label to contour map
}
