% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_contour.R
\name{plot3d}
\alias{plot3d}
\title{This function plots the fitted 3d surface for the expression value of a gene}
\usage{
plot3d(toPlot = list(x, y, z), DF, logScale = FALSE)
}
\arguments{
\item{toPlot}{is a list containing the predicted value for the x, y and z axis}

\item{DF}{is the data frame containing the information for the samples used in the fitting process}
}
\value{
a plotly object
}
\description{
This function plots the fitted 3d surface for the expression value of a gene
}
