% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hls_genes_clustering.R
\name{plot_clusters_prototypes}
\alias{plot_clusters_prototypes}
\title{this function plots the clusters prototype}
\usage{
plot_clusters_prototypes(meanXYZ, nR = 2, contour_size = 0.05)
}
\arguments{
\item{meanXYZ}{a list the contour object for each prototype computed with the function create_prototypes}

\item{nR}{the number of rows to use in the plot}

\item{contour_size}{number specifying the distance between each contour line}
}
\description{
this function plots the clusters prototype
}
