% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hls_genes_clustering.R
\name{plot_clusters_prototypes}
\alias{plot_clusters_prototypes}
\title{this function plots the clusters prototype}
\usage{
plot_clusters_prototypes(meanXYZ, nR = 2, contour_size = 0.05)
}
\arguments{
\item{meanXYZ}{a list the contour object for each prototype computed with the function create_prototypes}

\item{nR}{the number of rows to use in the plot}

\item{contour_size}{number specifying the distance between each contour line}
}
\description{
this function plots the clusters prototype
}
\examples{
data("FC_WY14643")
exp_data = fc_data
pheno_data = pdata
  PvalMat = compute_anova_dose_time(exp_data, pheno_data,dose_index = 2,time_point_index = 3)
  ItemsList = build_items_list(PvalMat)
  responsive_genes = unique(c(unlist(ItemsList$Dose),unlist(ItemsList$Time),
                              unlist(ItemsList$`Dose:Time:DoseTime`),
                              unlist(ItemsList$`Dose:Time`)))
  contour_res = create_contour(exp_data, pheno_data, responsive_genes,
                              dose_index = 2,time_point_index =3 ,gridSize = 50)
  hls_res = hls_genes_clustering(contour_res$GenesMap,  nClust = c(5,10,15,20,25),
                                 method="pearson", hls.method = "ward")

  clpr = create_prototypes(clust_res = hls_res,summaryMat = hls_res$summaryMat,contour_res )
  plot_clusters_prototypes(clpr$meanXYZ, nR = 2)

}
