% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{write_xlsx_for_funmappone}
\alias{write_xlsx_for_funmappone}
\title{This function create an excel file with the same format of the input need by the FunMappOne tool}
\usage{
write_xlsx_for_funmappone(optimal_clustering,
  filePath = "../contour_clustering/gene_clustering.xlsx")
}
\arguments{
\item{optimal_clustering}{is a numeric vector with the clustering result for every gene}

\item{filePath}{is a string specifying the path of the xlsx file}
}
\description{
This function create an excel file with the same format of the input need by the FunMappOne tool
}
